+++ .babel-plugin-macrosrc.js
@@ -1,0 +1,8 @@
+module.exports = {
+  styledComponents: {
+    ssr: true,
+    displayName: false,
+    fileName: false,
+    pure: true
+  }
+};