+++ .babelrc.js
@@ -7,6 +7,7 @@ module.exports = {
     require.resolve('@babel/preset-react')
   ],
   plugins: [
+    require.resolve('babel-plugin-macros'),
     require.resolve('@babel/plugin-external-helpers'),
     require.resolve('@babel/plugin-syntax-dynamic-import'),
     require.resolve('@babel/plugin-transform-runtime'),