+++ src/setupTests.ts
@@ -3,4 +3,3 @@ import Adapter from 'enzyme-adapter-react-16';
 import 'jsdom-global/register';
 
 configure({ adapter: new Adapter() });
-